refactor(format): extract csv export owner
